+++ tests/testthat/testthat.pager.R
@@ -1,6 +1,7 @@
 library(diffobj)
 context("pager")
 
+void <- function(x) NULL
 test_that("Specifying pager", {
   style <- gdo("diffobj.style")
   if(is.null(style)) style <- StyleAnsi8NeutralYb()
@@ -101,7 +102,8 @@ test_that("viewer vs browser", {
   )
 })
 test_that("blocking", {
-  # Note that readline just proceeds in non-interactive mode
+  # Note that readline just proceeds in non-interactive mode, which is why we
+  # need the mock here
 
   with_mock(
     "diffobj:::interactive"=function() FALSE,
@@ -118,16 +120,57 @@ test_that("blocking", {
           suppressWarnings(make_blocking(sum, invisible=FALSE)(1:10))
         )[['visible']]
       )
-    }
-  )
-})
+  })
+  with_mock(
+    "diffobj:::readline"=function(...) warning("readline"),
+    {
+      expect_warning(
+        show(
+          diffChr(
+            "a", "b", format='raw',
+            pager=list(pager=void, make.blocking=TRUE, threshold=0)
+          )
+        ),
+        "readline"
+      )
+      expect_warning(
+        show(
+          diffChr(
+            "a", "b", format='html',
+            pager=list(
+              pager=void, make.blocking=NA, threshold=0,
+              file.keep=FALSE
+            )
+        ) ),
+        "readline"
+      )
+      expect_warning(
+        show(
+          diffChr(
+            "a", "b", format='html',
+            pager=list(pager=void, file.keep=TRUE)
+        ) ),
+        NA
+      )
+      expect_warning(
+        show(
+          diffChr(
+            "a", "b", format='html',
+            pager=list(pager=void, make.blocking=NA, file.keep=TRUE)
+        ) ),
+        NA
+) } )})
 test_that("html page output", {
-  pager <- PagerBrowser(pager=function(x) cat(readLines(x), sep="\n"))
+  pager <- PagerBrowser(
+    pager=function(x) cat(readLines(x), sep="\n"), make.blocking=FALSE
+  )
   expect_equal(
     capture.output(show(diffChr("A", "B", pager=pager, style=StyleRaw()))),
     c("< \"A\"       > \"B\"     ", "@@ 1 @@     @@ 1 @@   ", "< A         > B       ")
   )
-  pager.warn <- PagerBrowser(pager=function(x) cat(readLines(x), sep="\n"))
+  pager.warn <- PagerBrowser(
+    pager=function(x) cat(readLines(x), sep="\n"), make.blocking=FALSE
+  )
   expect_error(
     diffChr(
       "A", "B", pager=pager.warn, format="html", style=list(js="notafile")
@@ -197,5 +240,24 @@ test_that("pager_is_less", {
     expect_false(pager_is_less())
   })
   expect_false(diffobj:::file_is_less(tempfile()))
+})
+test_that("file.keep", {
+  f <- tempfile()
+  show(
+    diffChr(
+      "A", "B", format='raw',
+      pager=list(pager=void, file.keep=TRUE, file.path=f, threshold=0L)
+  ) )
+  expect_equal(
+    readLines(f),
+    c("< \"A\"       > \"B\"     ", "@@ 1 @@     @@ 1 @@   ",
+      "< A         > B       ")
+  )
+  show(
+    diffChr(
+      "A", "B", format='raw',
+      pager=list(pager=void, file.keep=FALSE, file.path=f, threshold=0L)
+  ) )
+  expect_false(file.exists(f))
 })