+++ NEWS.md
@@ -1,7 +1,11 @@
 # diffobj
 
-## v0.1.12
+## v0.2.0
 
+* [#129](https://github.com/brodieG/diffobj/issues/129): Allow pager
+  specification via lists rather than full `Pager` objects for easier changes to
+  defaults.  As part of this we changed `StyleRaw` objects to use default
+  pager instead of `PagerOff`.
 * [#127](https://github.com/brodieG/diffobj/issues/127): Width CSS conflicts
   with bootstrap (reported by @eckyu, debugged/fixed by @cpsievert).
 * [#126](https://github.com/brodieG/diffobj/issues/126): Add embedding diffs in